# Improve HasEntProp performance