@asherkin 2023-01-18
Nice work, thanks for this! I'll try and take a look in the next few days - prod me on Discord if I forget.