@Kenzzer 2023-01-22
Can a method be exposed for extensions to purge the cache ?
https://canary.discord.com/channels/335290997317697536/335290997317697536/1034852606608166942